Create README.md
As a professional Graphics Designer Team, we would like to offer you:

1.	Clipping Path
2.	Multi-Clipping Path
3.	Image Manipulation
4.	Masking
5.	Retouching
6.	Shadow

As a professional Social Media Manager Team, we would like to offer you:

•	We can create whatever you need to make your Social Media Profiles shine like Facebook, Linkedin, Instagram, Twitter, Google +, Pinterest and more if you have any special requirements.   
•	We can work with the Existing Content that you have or I can create a new one. 
•	Increase website traffic
•	Page setup
•	Optimization
•	Post creation ( with your logo included)
•	HashTag Research
•	 Content Creation
•	 Design 
•	 Monthly Reporting

We will provide you the best efforts to send the visitors to your site and increase your social media followers a well. We will apply the latest marketing tools and techniques to promote your account's visibility and differentiate as to reach it in the highest pick among your competitors.

Feel free to message me on Fiverr.